refactor: pr 수정
리뷰 점수 double->int
입력 예외 NOT_FOUND->BAD_REQUEST
제목 길이 10에서 30으로 수정
리뷰 삭제할 때 이미지 삭제도 추가

+++ src/main/kotlin/sosteam/deamhome/domain/review/entity/Review.kt
@@ -15,7 +15,7 @@ import sosteam.deamhome.global.image.entity.Image
 class Review(
 	var title: String,
 	var content: String,
-	var score: Double = 0.0,
+	var score: Int = 0,
 	var status: Boolean = false,
 	val userId: String,
 	val itemId: String,
